+++ internal/pkg/service/common/task/orchestrator/doc.go
@@ -1,0 +1,3 @@
+// Package orchestrator combines the distribution.Node and the task.Node,
+// to run a task only on one node in the cluster, as a reaction to a watch event.
+package orchestrator
