+++ internal/pkg/service/common/task/orchestrator/orchestrator.go
@@ -1,5 +1,3 @@
-// Package orchestrator combines the distribution.Node and the task.Node,
-// to run a task only on one node in the cluster, as a reaction to a watch event.
 package orchestrator
 
 import (
@@ -23,83 +21,6 @@ const (
 	SpanNamePrefix = "keboola.go.buffer.orchestrator"
 )
 
-// Config configures the orchestrator.
-//
-// The orchestrator watches for all Source.WatchEvents in the etcd Source.WatchPrefix.
-// For each event, the TaskFactory is invoked.
-// If you want to handle only an event type, use StartTaskIf, see below.
-//
-// The TaskFactory is invoked only if the generated DistributionKey is assigned to the orchestrator.Node.
-// All events are received by all nodes, but each node decides
-// whether the DistributionKey is assigned to it and task should be started or not.
-//
-// Each orchestrator.Node in the cluster contains the distribution.Node instance
-// Nodes are synchronized, so they make the same decisions.
-// Decision is made by the distribution.Assigner.
-// Any short-term difference can lead to task duplication, but it can be prevented by the task locks.
-//
-// The watched Source.WatchPrefix is rescanned at Source.ReSyncInterval and when there is any change in workers distribution.
-// On the rescan, the existing tasks are NOT cancelled.
-// Task duplication is prevented by the task locks.
-//
-// The TaskFactory may return nil, then the event will be ignored.
-type Config[T any] struct {
-	// Name is orchestrator name used by logger.
-	Name string
-	// Source triggers new tasks.
-	Source Source[T]
-	// DistributionKey determines which worker processes the task. See distribution package.
-	DistributionKey func(event etcdop.WatchEventT[T]) string
-	// Lock, only one task with the lock can run at a time.
-	// If it is not set, the TaskKey is used as the lock name.
-	Lock func(event etcdop.WatchEventT[T]) string
-	// TaskKey defines etcd prefix where the task will be stored in etcd.
-	// CreatedAt datetime and a random suffix are always appended to the TaskID.
-	TaskKey func(event etcdop.WatchEventT[T]) task.Key
-	// StartTaskIf, if set, it determines whether the task is started or not
-	StartTaskIf func(event etcdop.WatchEventT[T]) (skipReason string, start bool)
-	// TaskCtx must return a task context with a deadline.
-	TaskCtx task.ContextFactory
-	// TaskFactory is a function that converts an etcd watch event to a task.
-	TaskFactory TaskFactory[T]
-}
-
-func (c Config[T]) Validate() error {
-	errs := errors.NewMultiError()
-	if c.Name == "" {
-		errs.Append(errors.New("orchestrator name must be configured"))
-	}
-	if c.Source.WatchPrefix.Prefix() == "" {
-		errs.Append(errors.New("source watch prefix definition must be configured"))
-	}
-	if c.Source.RestartInterval <= 0 {
-		errs.Append(errors.New("restart interval must be configured"))
-	}
-	if c.DistributionKey == nil {
-		errs.Append(errors.New("task distribution factory key factory must be configured"))
-	}
-	if c.TaskKey == nil {
-		errs.Append(errors.New("task key must be configured"))
-	}
-	if c.TaskCtx == nil {
-		errs.Append(errors.New("task ctx factory must be configured"))
-	}
-	if c.TaskFactory == nil {
-		errs.Append(errors.New("task factory must be configured"))
-	}
-	return errs.ErrorOrNil()
-}
-
-type Source[T any] struct {
-	// WatchPrefix defines an etcd prefix that is watched by GetAllAndWatch.
-	// Each event triggers new task.
-	WatchPrefix etcdop.PrefixT[T]
-	// WatchEtcdOps contains additional options for the watch operation
-	WatchEtcdOps []etcd.OpOption
-	// ReSyncInterval defines the interval after all keys in the prefix are processed again.
-	RestartInterval time.Duration
-}
-
 // Orchestrator creates a task for each watch event, but only on one worker node in the cluster.
 // Decision is made by the distribution.Assigner.
 // See documentation of: distribution.Node, task.Node, Config[R].
@@ -112,8 +33,6 @@ type orchestrator[T any] struct {
 	tasks  *task.Node
 	config Config[T]
 }
-
-type TaskFactory[T any] func(event etcdop.WatchEventT[T]) task.Fn
 
 type dependencies interface {
 	Clock() clock.Clock
@@ -129,7 +48,7 @@ func Start[T any](ctx context.Context, wg *sync.WaitGroup, d dependencies, confi
 		panic(err)
 	}
 
-	w := &orchestrator[T]{
+	o := &orchestrator[T]{
 		clock:  d.Clock(),
 		logger: d.Logger().AddPrefix(fmt.Sprintf("[orchestrator][%s]", config.Name)),
 		tracer: d.Telemetry().Tracer(),
@@ -140,9 +59,9 @@ func Start[T any](ctx context.Context, wg *sync.WaitGroup, d dependencies, confi
 	}
 
 	// Delete events are not needed/ignored
-	w.config.Source.WatchEtcdOps = append(w.config.Source.WatchEtcdOps, etcd.WithFilterDelete())
+	o.config.Source.WatchEtcdOps = append(o.config.Source.WatchEtcdOps, etcd.WithFilterDelete())
 
-	return w.start(ctx, wg)
+	return o.start(ctx, wg)
 }
 
 func (o orchestrator[R]) start(ctx context.Context, wg *sync.WaitGroup) <-chan error {
